+++ lib/apollovm.dart
@@ -28,6 +28,8 @@ export 'src/languages/csharp/csharp_parser.dart';
 export 'src/languages/csharp/csharp_runner.dart';
 export 'src/languages/dart/dart_parser.dart';
 export 'src/languages/dart/dart_runner.dart';
+export 'src/languages/go/go_parser.dart';
+export 'src/languages/go/go_runner.dart';
 export 'src/languages/java/java11/java11_parser.dart';
 export 'src/languages/java/java11/java11_runner.dart';
 export 'src/languages/wasm/wasm_parser.dart';