+++ test/languages/apollovm_languages_extended_test.dart
@@ -8,6 +8,7 @@
   'python',
   'kotlin',
   'lua',
+  'go',
 ])
 library;